tech(api): move method from campaign participation repository in campaign context to campaign-participation context
